move anagram types
