+++ src/services/anagram.service.ts
@@ -1,7 +1,8 @@
 import { AxiosError } from 'axios';
 import client from './services.client';
 import { apiRoutes } from '../utils/constants.util';
-import { Anagram, ApiResponse } from '../types/types';
+import { Anagram } from '../types/anagram.interface';
+import { ApiResponse } from '../types/types';
 
 // TODO: log error to DB
 const AnagramService = {
