+++ src/types/anagram.interface.ts
@@ -1,0 +1,18 @@
+export interface BaseAnagram {
+  id: number,
+  anagram: string,
+  level_word: string
+}
+
+export interface Anagram extends BaseAnagram {
+  solved: boolean
+}
+
+export interface AnagramsHashMap {
+  [key: string]: {
+    [key: string]: Anagram
+  }
+}
+
+// export interface AnagramProps {
+// }
